backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
neo4j==5.28.1
python-dotenv==1.1.0
